requirements.txt
fastapi
uvicorn[standard]
docker
PyYAML